Google Font Added : Roboto
Since NextJS  don't have a public/index.html file to setup custom fonts we have create a file as pages/_document.tsx and then the configurations we've applied will be read by the application. This file works just like a index.html in a traditional cra project.

After that , we just have to pass the font-family : 'Roboto' property to our home.module.css to signify to the app to use the font we've 'imported'