Global Styles
First Global Styles set. Just added some more default styles and some media queries to deal with different devices screen sizes since we're using rem units for the font-size. So the accessibility of the app is covered in case of users with some disabilities need to scale up the fonts of the application.