FaunaDB - Install & Initial Setup
Installed FaundDB since we're dealing with a Server-less application. After setting up an account in FaunaDB, and generate a key for the app, we can configure the integration between app and database. Simply by creating a fauna.ts file inside of services folder, and configure it to receive the related keys.