Posts Page & Individual Post Page
Posts Page is almost complete, need only to implement the verification for only users to be able to read the full content of the posts, since it will be for members only the full article.

Finished implementing the dynamic routing from Posts to each post using the [slug].tsx component to better organize the application.

Added also the first styles for the individual post.