Post Preview - Part 2
Now the post preview is complete and redirecting users that are not logged to the Home page where they can subscribe and have access to the content. Also automatically will be checking if the user is already subscribed and if it is, the post will be displayed in full length.